Added improved controller navigation.
Changed quite a bit of the code to change controller navigation. Now pressing left and right goes through the images. There was a table of bytes added for the palette and map data, allowing the tile bank to be matched up with its corresponding data in the other banks. Code is a bit of a mess but fairly well commented I hope.